# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "Food-Delivery",
     "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
     
       "version": "1.0"
   },
   "produces": ["application/json"],
   "paths": {
...
    "/api/mobile/company/getCompanies": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/company/getCompanies",
        "operationId": "index",
        "tags": ["All Vendors listing"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "deliveryType",  "in": "query",  "required":"true","type":"string"},
          { "name": "page",  "in": "query","type":"string"},
          { "name": "limit",  "in": "query","type":"string"},
          { "name": "lat",  "in": "query","type":"string"},
          { "name": "lng",  "in": "query","type":"string"},
          { "name": "itemType",  "in": "query","type":"string"}



     

     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/company/detail": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/company/detail",
        "operationId": "index",
        "tags": ["Get Vendor Detail"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "companyId",  "in": "query",  "required":"true","type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/company/shipmentCharges": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/company/shipmentCharges",
        "operationId": "index",
        "tags": ["Get Shipment Charges"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":"true","type":"string"},
          { "name": "addressId",  "in": "formData",  "required":"true","type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/cart/list": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/cart/list",
        "operationId": "index",
        "tags": ["Cart Listing"],
        "description": "get cart listing",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/cart/add": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/cart/add",
        "operationId": "index",
        "tags": ["Add to Cart"],
        "description": "Add Item to Cart",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "deliveryType",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":true,"type":"string"}


     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/cart/update": {
      "put": {
        "x-swagger-router-controller": "/api/mobile/cart/update",
        "operationId": "index",
        "tags": ["Update Cart"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "cartId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "addressId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "companyId",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
          { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"}
          

     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/cart/remove": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/cart/remove",
        "operationId": "index",
        "tags": ["Remove Item from Cart"],
        "description": "get Prduct listing by Category ID",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true"},
          { "name": "cartId",  "in": "query",  "required":"true"}


     ],

  
        "responses": {}
      }
    },
    "/api/mobile/cart/clear": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/cart/clear",
        "operationId": "index",
        "tags": ["Clear Cart"],
        "description": "get Prduct listing by Category ID",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true"}


     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/notification/": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/notification/",
        "operationId": "index",
        "tags": ["Notification Listing"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"}
         ],
        "responses": {}
      }
    },
    "/api/mobile/notification/clearAll": {
      "delete": {
        "x-swagger-router-controller": "/api/mobile/notification/clearAll",
        "operationId": "index",
        "tags": ["Clear All Notifications"],
        "description": "",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"}
         ],
        "responses": {}
      }
    }
   }
}